+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+* The destination directory of iai callgrind output files changes from
+`/workspace/$CARGO_PKG_NAME/target/iai` to `/workspace/target/iai/$CARGO_PKG_NAME` and respects the
+`CARGO_TARGET_DIR` environment variable
+
+### [0.6.0] - 2023-08-20
+
 ### Added
 
 * (#3) builder api for binary benchmarks